First Commit - Good Start
Commiting the test code for the Unity project.  Import this project within unity.  To try this yourself, within unity click the project tab (its near the bottom but if its not there you can open up the window).  There will be an assets folder with everything inside assuming you loaded in this project properly.  You should be able to do a pull within unity assuming you want to start with a blank project as well, but I am not sure.  In the assets folder, load testscene and everything will load in as I have configured.  Additionally, the code is in TerrainGen.cs and some helpful links with comments are in links.txt.  Again, both are in the asset folder.  I will go into further details in person for the Unity stuff, like how to add code to RandTerrain, loading in sprites, backgrounds, etc.  I'm no expert and tutorials will go a long way as opposed to what I say.  Lets get after it!